// js/stat-evaluation-player/src/generated/BoostStateEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type BoostStateEvent = { frame: number, time: number, player_id: RemoteIdTs, is_team_0: boolean, boost_amount: number, boost_before: number | null, };